+++ module-apps/application-antenna/windows/ScanModesWindow.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2020, Mudita Sp. z.o.o. All rights reserved.
+﻿// Copyright (c) 2017-2020, Mudita Sp. z.o.o. All rights reserved.
 // For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
 
 #include "ScanModesWindow.hpp"
@@ -12,7 +12,7 @@
 
 #include "../AntennaAppStyle.hpp"
 #include "../ApplicationAntenna.hpp"
-#include "service-cellular/api/CellularServiceAPI.hpp"
+#include <service-cellular/CellularServiceAPI.hpp>
 
 namespace gui
 {