+++ module-apps/application-call/ApplicationCall.hpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2020, Mudita Sp. z.o.o. All rights reserved.
+﻿// Copyright (c) 2017-2020, Mudita Sp. z.o.o. All rights reserved.
 // For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
 
 #pragma once
@@ -7,7 +7,7 @@
 #include "Service/Message.hpp"
 #include "Service/Timer.hpp"
 #include "SystemManager/SystemManager.hpp"
-#include <service-cellular/api/CellularServiceAPI.hpp>
+#include <service-cellular/CellularMessage.hpp>
 #include <time/time_conversion.hpp>
 
 namespace app